Add auth middleware
Added verifyAdminToken middleware which determines if the the JWTs role is admin. Added addRole middleware which adds the id and role to the local response variables, this is used in the logic for endpoints that need make decisions based on role and id such as the user show and update routes.